implementando proyecto a fragments
