+++ app/src/main/assets/mock_ciudades.json
@@ -43,7 +43,7 @@
     "titulo": "Playa Burj Al Arab",
     "descripcion": "Si lo que quieres es relajarte en los días soleados, la Burj Al Arab es perfecta para ti. Ubicada en Jumeirah, se divide en dos partes: una para los surfistas y otra para los turistas que desean relajarse. En ésta última podrás encontrar servicios de duchas y vestuarios a la vez que obtienes una vista espectacular del Hotel Burj Al Arab, por eso se ha convertido en uno de los lugares atractivos en Dubái.",
     "descripcion_corta": "Incursiona en el surf con las olas de la Playa Burj Al Arab.",
-    "ubicacion": "Se ubicda a orillas de la ciudad",
+    "ubicacion": "Se ubica a orillas de la ciudad",
     "tiempo": "30 °C, viento del 10 a 20 km/h, humedad del 65 %",
     "sitios": "- Burj Al-Arab Hotel \n- Torre Burj khalifa\n-  Isla artificial \n- Playa de Jumeira",
     "ruta_imagen": "https://p0.pikist.com/photos/393/564/dubai-sail-uae-burj-al-arab-hotel-building-beach-canvas-jumeirah.jpg",
