+++ app/src/main/res/layout/fragment_list_.xml
@@ -46,16 +46,6 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent" />
 
-        <ImageButton
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginEnd="4dp"
-            android:background="@color/alfa"
-            android:src="@drawable/menu_puntos_negro"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintBottom_toBottomOf="@id/tvDubai"
-            app:layout_constraintTop_toTopOf="@id/tvDubai" />
-
         <TextView
             android:id="@+id/textView"
             android:layout_width="match_parent"
@@ -78,6 +68,17 @@
             app:layout_constraintTop_toBottomOf="@id/tvDubai"
             tools:ignore="MissingConstraints"
             tools:layout_editor_absoluteX="1dp" />
+
+        <ImageButton
+            android:id="@+id/imageButton"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:clickable="true"
+            android:src="@drawable/menu_puntos_negro"
+            tools:ignore="MissingConstraints"
+            tools:layout_editor_absoluteX="363dp"
+            tools:layout_editor_absoluteY="300dp"
+            />
 
 
     </androidx.constraintlayout.widget.ConstraintLayout>
